Reorganize logic
